main logic implemented

+++ src/main/java/com/miro/widgetservice/service/impl/WidgetServiceImpl.java
@@ -3,17 +3,18 @@ package com.miro.widgetservice.service.impl;
 import com.miro.widgetservice.converter.WidgetConverter;
 import com.miro.widgetservice.dto.WidgetDto;
 import com.miro.widgetservice.exception.WidgetServiceException;
-import com.miro.widgetservice.model.WidgetInMemory;
-import com.miro.widgetservice.repository.IdGenerator;
+import com.miro.widgetservice.model.Widget;
 import com.miro.widgetservice.repository.WidgetRepository;
 import com.miro.widgetservice.service.WidgetService;
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import static com.miro.widgetservice.repository.WidgetMemoryRepository.getWidgetDatasource;
-
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class WidgetServiceImpl implements WidgetService {
@@ -22,43 +23,25 @@ public class WidgetServiceImpl implements WidgetService {
 
     private final WidgetConverter widgetConverter;
 
-    private final IdGenerator idGenerator;
-
     @Override
     public WidgetDto create(WidgetDto widgetDto) {
+        log.info("Create new widget {}", widgetDto);
+        Widget widgetToUpdate = widgetConverter.convert(widgetDto);
+        widgetToUpdate.setModificationDate(LocalDateTime.now());
+        Widget savedWidget = widgetRepository.save(widgetToUpdate);
 
-        WidgetInMemory widget = widgetConverter.convert(widgetDto);
-
-        WidgetInMemory byZIndex = widgetRepository.findByZIndex(widget.getZIndex());
-        if (byZIndex == null) {
-            long id = idGenerator.generateNextId();
-            widget.setId(id);
-            getWidgetDatasource().put(id, widget);
-        } else {
-
-            WidgetInMemory existByZ = getWidgetDatasource().values()
-                .stream()
-                .filter(w -> w.getZIndex().equals(widget.getZIndex()))
-                .findFirst()
-                .orElseThrow();
-
-            getWidgetDatasource().values().stream().reduce(existByZ, (l, r) -> {
-
-                return l;
-            });
-
-        }
-        return null;
+        return widgetConverter.convert(savedWidget);
     }
 
     @Override
     public WidgetDto update(Long id, WidgetDto widgetDto) {
         widgetDto.setId(id);
-
+        log.info("Update widget {}", widgetDto);
         if (isExist(id)) {
-            WidgetInMemory widget = widgetConverter.convert(widgetDto);
-            widgetRepository.save(widget);
-            return widgetDto;
+            Widget widget = widgetConverter.convert(widgetDto);
+            widget.setModificationDate(LocalDateTime.now());
+            Widget savedWidget = widgetRepository.save(widget);
+            return widgetConverter.convert(savedWidget);
         }
 
         throw new WidgetServiceException("Widget with id " + id + " does not exist");
@@ -66,23 +49,45 @@ public class WidgetServiceImpl implements WidgetService {
 
     @Override
     public WidgetDto findById(Long id) {
-        WidgetInMemory savedWidget = widgetRepository.findById(id).orElseThrow(() -> new WidgetServiceException("Widget does not exist"));
-        return widgetConverter.convert(savedWidget);
+        log.info("Find widget by id {}", id);
+
+        Widget existWidget = widgetRepository.findById(id)
+            .orElseThrow(() -> new WidgetServiceException("Widget with id " + id + " does not exist"));
+
+        return widgetConverter.convert(existWidget);
     }
 
     @Override
-    public List<WidgetDto> findAll() {
-        return widgetRepository.findAll().stream()
+    public List<WidgetDto> findAll(Integer page) {
+        log.info("Find all widgets. Page: {}", page);
+
+        if (Objects.isNull(page)) {
+            return widgetRepository.findAll().stream()
+                .map(widgetConverter::convert)
+                .collect(Collectors.toList());
+        }
+        if (page < 0) {
+            throw new WidgetServiceException("Page cannot be negative");
+        }
+        return widgetRepository.findAll(page).stream()
             .map(widgetConverter::convert)
             .collect(Collectors.toList());
     }
 
     @Override
     public void deleteById(Long id) {
-        widgetRepository.deleteById(id);
+        log.info("Delete widget by id {}", id);
+
+        if (isExist(id)) {
+            widgetRepository.deleteById(id);
+            return;
+        }
+        throw new WidgetServiceException("Widget with id " + id + " does not exist");
     }
 
     private boolean isExist(Long id) {
+        log.info("Widget is exist by id {}", id);
+
         return widgetRepository.isExist(id);
     }
 }
